Various fixes and changes
1. Add setSnapAtPosition to SwipeListView.java.
This changes the swipe-mode: It will callback the onPassSnapPosition instead of deleting/revealing/checking, alerting the listener when a the user has swiped at least certain distance (deltaX). The unit is in pixels.

2. Added onPassSnapPosition.
The 'after' boolean is true when the user has swiped passed the 'snapAtPosition'.
The 'after' boolean is false when the user has not (yet) swiped passed the 'snapAtPosition'.
The 'executed' boolean is true only when the user had let go of the item being swiped, i.e. the user ended the swipe and the list-item snaps back.

3. Make animation/moving/slding smoother when the user's drag goes beyond the 'slop' value.